+++ src/stapi_fastapi/backends/product_backend.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 from typing import Protocol
 
 from fastapi import Request
+from returns.result import ResultE
 
 from stapi_fastapi.models.opportunity import Opportunity, OpportunityRequest
 from stapi_fastapi.models.order import Order, OrderRequest
@@ -15,25 +16,23 @@ class ProductBackend(Protocol):  # pragma: nocover
         product_router: ProductRouter,
         search: OpportunityRequest,
         request: Request,
-    ) -> list[Opportunity]:
+    ) -> ResultE[list[Opportunity]]:
         """
         Search for ordering opportunities for the  given search parameters.
 
-        Backends must validate search constraints and raise
+        Backends must validate search constraints and return
         `stapi_fastapi.exceptions.ConstraintsException` if not valid.
         """
-        ...
 
     async def create_order(
         self,
         product_router: ProductRouter,
         search: OrderRequest,
         request: Request,
-    ) -> Order:
+    ) -> ResultE[Order]:
         """
         Create a new order.
 
-        Backends must validate order payload and raise
+        Backends must validate order payload and return
         `stapi_fastapi.exceptions.ConstraintsException` if not valid.
         """
-        ...