+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [unreleased]
 
 ### Added
 
@@ -13,6 +13,35 @@ none
 
 ### Changed
 
+- OrderStatusCode and ProviderRole are now StrEnum instead of (str, Enum)
+- All types using `Result[A, Exception]` have been replace with the equivalent type `ResultE[A]`
+
+### Deprecated
+
+none
+
+### Removed
+
+none
+
+### Fixed
+
+none
+
+### Security
+
+none
+
+## [v0.2.0] - 2024-11-23
+
+### Added
+
+none
+
+### Changed
+
+- RootBackend and ProductBackend protocols use `returns` library types Result and Maybe instead of
+  raising exceptions.
 - Create Order endpoint from `.../order` to `.../orders`
 - Order field `id` must be a string, instead of previously allowing int. This is because while an
   order ID may an integral numeric value, it is not a "number" in the sense that math will be performed
@@ -50,5 +79,6 @@ Initial release
 - Add links `opportunities` and `create-order` to Product
 - Add link `create-order` to OpportunityCollection
 
-[unreleased]: https://github.com/stapi-spec/stapi-fastapi/compare/v0.1.0...main
+[unreleased]: https://github.com/stapi-spec/stapi-fastapi/compare/v0.2.0...main
+[v0.2.0]: https://github.com/stapi-spec/stapi-fastapi/tree/v0.2.0
 [v0.1.0]: https://github.com/stapi-spec/stapi-fastapi/tree/v0.1.0